restructure identity files
