+++ src/ui/identities/mod.rs
@@ -1,4 +1,6 @@
 pub mod add_existing_identity_screen;
 pub mod add_new_identity_screen;
 pub mod identities_screen;
+pub mod keys;
 pub mod register_dpns_name_screen;
+pub mod withdraw_from_identity_screen;
